test(#39609): move from e2e to intergration
