+++ test/integration/next-dynamic-with-suspense/pages/thing.js
@@ -1,0 +1,3 @@
+export default function Thing() {
+  return 'Thing'
+}
